refactor unit test project
