+++ azure-quantum/tests/integration/test_streaming_problem.py
@@ -19,7 +19,7 @@ from typing import List
 from azure.quantum import Workspace
 from azure.quantum.optimization import StreamingProblem, Problem, ProblemType, Term
 from azure.quantum.storage import download_blob
-from integration_test_util import create_workspace
+from ..testutils.integration_test_util import create_workspace
 
 class TestStreamingProblem(unittest.TestCase):
     def __test_upload_problem(
